Refactor: move tests for `ParsedType` into `tests/test_typing.py`
